try unifying tables
